one more correction to track-hit-index assn

+++ TrackFinder/PMAlgTrackMaker_module.cc
@@ -219,7 +219,7 @@ PMAlgTrackMaker::PMAlgTrackMaker(fhicl::ParameterSet const & p) :
 	this->reconfigure(p);
 	produces< std::vector<recob::Track> >();
 	produces< std::vector<recob::SpacePoint> >();
-	produces< art::Assns<recob::Track, recob::Hit, size_t> >();
+	produces< art::Assns<recob::Track, recob::Hit, unsigned int> >();
 	produces< art::Assns<recob::Track, recob::SpacePoint> >();
 	produces< art::Assns<recob::SpacePoint, recob::Hit> >();
 }
@@ -861,7 +861,7 @@ void PMAlgTrackMaker::produce(art::Event& evt)
 	std::unique_ptr< std::vector< recob::Track > > tracks(new std::vector< recob::Track >);
 	std::unique_ptr< std::vector< recob::SpacePoint > > allsp(new std::vector< recob::SpacePoint >);
 
-	std::unique_ptr< art::Assns< recob::Track, recob::Hit, size_t > > trk2hit(new art::Assns< recob::Track, recob::Hit, size_t >);
+	std::unique_ptr< art::Assns< recob::Track, recob::Hit, unsigned int > > trk2hit(new art::Assns< recob::Track, recob::Hit, unsigned int >);
 	std::unique_ptr< art::Assns< recob::Track, recob::SpacePoint > > trk2sp(new art::Assns< recob::Track, recob::SpacePoint >);
 	std::unique_ptr< art::Assns< recob::SpacePoint, recob::Hit > > sp2hit(new art::Assns< recob::SpacePoint, recob::Hit >);
 
@@ -955,7 +955,7 @@ void PMAlgTrackMaker::produce(art::Event& evt)
 					size_t trkIdx = tracks->size() - 1;
 					art::ProductID trkId = getProductID< std::vector<recob::Track> >(evt);
 					art::Ptr<recob::Track> trkPtr(trkId, trkIdx, evt.productGetter(trkId));
-					for (size_t hidx = 0; hidx < hits2d.size(); hidx++)
+					for (unsigned int hidx = 0; hidx < hits2d.size(); hidx++)
 						trk2hit->addSingle(trkPtr, hits2d[hidx], hidx);
 
 					util::CreateAssn(*this, evt, *tracks, *allsp, *trk2sp, spStart, spEnd);
